Open agreement contents in a popup window
