+++ android/source/VideoLocker/res/layout/activity_register.xml
@@ -108,9 +108,7 @@
                 android:id="@+id/layout_agreement"
                 android:layout_width="match_parent"
                 android:layout_height="wrap_content"
-                android:gravity="center_horizontal"
-                android:orientation="vertical"
-                android:visibility="gone" />
+                android:orientation="vertical"/>
 
             <org.edx.mobile.view.custom.ETextView
                 android:id="@+id/end_user_agreement_tv"
