+++ android/source/VideoLocker/res/values-en/strings.xml
@@ -183,5 +183,5 @@
 
     <string name="enrollment_failure">We could not enroll you in this course at this time.</string>
     <string name="try_again">TRY AGAIN</string>
-    <string name="you_are_now_enrolled">You are now enrolled</string>
+    <string name="you_are_now_enrolled">You have enrolled in this course</string>
 </resources>
