+++ android/source/VideoLocker/src/org/edx/mobile/base/BaseFragmentActivity.java
@@ -683,16 +683,22 @@ public class BaseFragmentActivity extends FragmentActivity implements NetworkSub
         getWindow().clearFlags(WindowManager.LayoutParams.FLAG_NOT_TOUCHABLE);
     }
 
-    public void showWebDialog(String fileName, boolean showTitle, String dialogTitle){
+    /**
+     * Displays a dialog which has a WebView container to display contents of given URL.
+     * @param url String
+     * @param showTitle
+     * @param dialogTitle
+     */
+    public void showWebDialog(String url, boolean showTitle, String dialogTitle) {
         //Show the dialog only if the activity is started. This is to avoid Illegal state
         //exceptions if the dialog fragment tries to show even if the application is not in foreground
         if(isActivityStarted()){
             WebViewDialogFragment webViewFragment = new WebViewDialogFragment();
-            webViewFragment.setDialogContents(fileName, showTitle, dialogTitle);
+            webViewFragment.setDialogContents(url, showTitle, dialogTitle);
             webViewFragment.setStyle(DialogFragment.STYLE_NORMAL,
                     android.R.style.Theme_Black_NoTitleBar_Fullscreen);
             webViewFragment.setCancelable(false);
-            webViewFragment.show(getSupportFragmentManager(), "dialog");
+            webViewFragment.show(getSupportFragmentManager(), "web-view-dialog");
         }
     }
