+++ android/source/VideoLocker/src/org/edx/mobile/module/registration/view/RegistrationEditTextView.java
@@ -1,8 +1,8 @@
-package org.edx.mobile.view.registration;
+package org.edx.mobile.module.registration.view;
 
+import android.text.InputFilter;
 import android.view.View;
-import android.widget.ArrayAdapter;
-import android.widget.Spinner;
+import android.widget.EditText;
 import android.widget.TextView;
 
 import com.google.gson.JsonElement;
@@ -10,41 +10,43 @@ import com.google.gson.JsonPrimitive;
 
 import org.edx.mobile.R;
 import org.edx.mobile.logger.Logger;
-import org.edx.mobile.model.registration.RegistrationFormField;
-import org.edx.mobile.model.registration.RegistrationOption;
+import org.edx.mobile.module.registration.model.RegistrationFormField;
 
-class RegistrationSpinnerView implements IRegistrationFieldView {
+class RegistrationEditTextView implements IRegistrationFieldView {
 
     protected static final Logger logger = new Logger(RegistrationEditTextView.class);
-    private RegistrationFormField mField;
+    protected RegistrationFormField mField;
     private View mView;
-    protected Spinner mInputView;
+    protected EditText mInputView;
     private TextView mErrorView, mInstructionView;
+    private IActionListener actionListener;
 
-    public RegistrationSpinnerView(RegistrationFormField field, View view) {
+    public RegistrationEditTextView(RegistrationFormField field, View view) {
         // create and configure view and save it to an instance variable
         this.mField = field;
         this.mView = view;
 
-        this.mInputView = (Spinner) view.findViewById(R.id.input_spinner);
-        this.mErrorView = (TextView) view.findViewById(R.id.input_spinner_error);
-        this.mInstructionView = (TextView) view.findViewById(R.id.input_spinner_instruction);
+        this.mInputView = (EditText) view.findViewById(R.id.txt_input);
+        this.mErrorView = (TextView) view.findViewById(R.id.txt_input_error);
+        this.mInstructionView = (TextView) view.findViewById(R.id.txt_input_instruction);
+
+        // set max lines for this input to be 1
+        mInputView.setLines(1);
+
+        // apply max length
+        if (mField.getRestriction().getMaxLength() > 0) {
+            // otherwise, you may end up disabling the field
+
+            InputFilter[] FilterArray = new InputFilter[1];
+            FilterArray[0] = new InputFilter.LengthFilter(mField.getRestriction().getMaxLength());
+            mInputView.setFilters(FilterArray);
+        }
 
         // set hint
-        mInputView.setPrompt(mField.getLabel());
+        mInputView.setHint(mField.getLabel());
 
-        // display default value and set the entries
-        ArrayAdapter<RegistrationOption> adapter = new ArrayAdapter<>(mInputView.getContext(), R.layout.list_row_spinner);
-        int i=0, selectedIndex=0;
-        for (RegistrationOption option : mField.getOptions()) {
-            if (option.isDefaultValue()) {
-                selectedIndex = i;
-            }
-            adapter.add(option);
-            i++;
-        }
-        mInputView.setAdapter(adapter);
-        mInputView.setSelection(selectedIndex);
+        // display default value
+        mInputView.setText(mField.getDefaultValue());
 
         // display instructions if available
         if (mField.getInstructions() != null && !mField.getInstructions().isEmpty()) {
@@ -61,12 +63,12 @@ class RegistrationSpinnerView implements IRegistrationFieldView {
     @Override
     public JsonElement getCurrentValue() {
         // turn text view content into a JsonElement and return it
-        return new JsonPrimitive(mField.getOptions().get(mInputView.getSelectedItemPosition()).getValue());
+        return new JsonPrimitive(mInputView.getText().toString());
     }
 
     @Override
     public boolean hasValue() {
-        return !mField.getOptions().get(mInputView.getSelectedItemPosition()).getValue().isEmpty();
+        return !mInputView.getText().toString().isEmpty();
     }
 
     @Override
@@ -119,5 +121,10 @@ class RegistrationSpinnerView implements IRegistrationFieldView {
     @Override
     public void setEnabled(boolean enabled) {
         mInputView.setEnabled(enabled);
+    }
+
+    @Override
+    public void setActionListener(IActionListener actionListener) {
+        // no actions for this field
     }
 }
