+++ android/source/VideoLocker/src/org/edx/mobile/module/registration/view/RegistrationTextAreaView.java
@@ -1,10 +1,10 @@
-package org.edx.mobile.view.registration;
+package org.edx.mobile.module.registration.view;
 
 import android.text.InputType;
 import android.view.Gravity;
 import android.view.View;
 
-import org.edx.mobile.model.registration.RegistrationFormField;
+import org.edx.mobile.module.registration.model.RegistrationFormField;
 
 class RegistrationTextAreaView extends RegistrationEditTextView {
