+++ android/source/VideoLocker/src/org/edx/mobile/view/dialog/WebViewDialogFragment.java
@@ -2,28 +2,28 @@ package org.edx.mobile.view.dialog;
 
 import android.os.Bundle;
 import android.support.v4.app.DialogFragment;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.View.OnClickListener;
 import android.view.ViewGroup;
 import android.webkit.WebView;
-import android.webkit.WebViewClient;
 import android.widget.Button;
+import android.widget.ProgressBar;
 
 import org.edx.mobile.R;
 import org.edx.mobile.logger.Logger;
-import org.edx.mobile.util.BrowserUtil;
 import org.edx.mobile.view.custom.ETextView;
+import org.edx.mobile.view.custom.URLInterceptorWebViewClient;
 
 public class WebViewDialogFragment extends DialogFragment {
     private final Logger logger = new Logger(getClass().getName());
 
     private static final String TAG = WebViewDialogFragment.class.getCanonicalName();
 
-    String fileName;
-    boolean showTitle;
-    String dialogTitle;
+    private String fileName;
+    private boolean showTitle;
+    private String dialogTitle;
+    private ProgressBar progress;
     
     public WebViewDialogFragment() {
     }   
@@ -40,30 +40,44 @@ public class WebViewDialogFragment extends DialogFragment {
 
         View v = inflater.inflate(R.layout.fragment_web_dialog,
                 container, false);
+
+        progress = (ProgressBar) v.findViewById(R.id.progress);
+
         try{
             WebView webView = (WebView)v.findViewById(R.id.eula_webView);
-            webView.setWebViewClient(new WebViewClient(){
+            URLInterceptorWebViewClient client = new URLInterceptorWebViewClient(webView);
+            client.setPageStatusListener(new URLInterceptorWebViewClient.IPageStatusListener() {
+
                 @Override
-                public boolean shouldOverrideUrlLoading(final WebView view, final String url) {
-                    BrowserUtil.open(getActivity(), url);
-                    return true;
+                public void onPageStarted() {
+                    progress.setVisibility(View.VISIBLE);
+                }
+
+                @Override
+                public void onPageFinished() {
+                    progress.setVisibility(View.GONE);
+                }
+
+                @Override
+                public void onPageLoadError() {
+                    progress.setVisibility(View.GONE);
                 }
             });
+
             if(fileName!=null){
                 webView.loadUrl(fileName);
-                //eulaWeb.loadUrl("file:///android_asset/EULA.htm");
             }
             ETextView tv_dialog_title = (ETextView)v.findViewById(R.id.tv_dialog_title);
-            View view_seperator = (View)v.findViewById(R.id.view_seperator);
+            View viewSeperator = v.findViewById(R.id.view_seperator);
             if(showTitle){
                 tv_dialog_title.setVisibility(View.VISIBLE);
-                view_seperator.setVisibility(View.VISIBLE);
+                viewSeperator.setVisibility(View.VISIBLE);
                 if(dialogTitle!=null){
                     tv_dialog_title.setText(dialogTitle);
                 }
             }else{
                 tv_dialog_title.setVisibility(View.INVISIBLE);
-                view_seperator.setVisibility(View.INVISIBLE);
+                viewSeperator.setVisibility(View.INVISIBLE);
             }
         }catch(Exception e){
             logger.error(e);
